refactor: move structuredDataScript to shared library
