+++ apps/ff-site/src/app/blog/[slug]/page.tsx
@@ -1,6 +1,7 @@
 import { ArticleLayout } from '@filecoin-foundation/ui/Article/ArticleLayout'
 import { BlogPostHeader } from '@filecoin-foundation/ui/BlogPostHeader'
 import { PageLayout } from '@filecoin-foundation/ui/PageLayout'
+import { StructuredDataScript } from '@filecoin-foundation/ui/StructuredDataScript'
 
 import { type SlugParams } from '@/types/paramsTypes'
 
@@ -13,7 +14,6 @@ import { createMetadata } from '@/utils/createMetadata'
 
 import { MarkdownContent } from '@/components/MarkdownContent'
 import { ShareArticle } from '@/components/ShareArticle'
-import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import { getBlogPostData, getBlogPostsData } from '../utils/getBlogPostData'
