+++ apps/ff-site/src/app/digest/[slug]/page.tsx
@@ -1,6 +1,7 @@
 import { ArticleLayout } from '@filecoin-foundation/ui/Article/ArticleLayout'
 import { DigestArticleHeader } from '@filecoin-foundation/ui/DigestArticleHeader'
 import { PageLayout } from '@filecoin-foundation/ui/PageLayout'
+import { StructuredDataScript } from '@filecoin-foundation/ui/StructuredDataScript'
 
 import { type SlugParams } from '@/types/paramsTypes'
 
@@ -12,7 +13,6 @@ import { createMetadata } from '@/utils/createMetadata'
 
 import { MarkdownContent } from '@/components/MarkdownContent'
 import { ShareArticle } from '@/components/ShareArticle'
-import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import {
   getDigestArticleData,
