+++ apps/ff-site/src/app/ecosystem-explorer/page.tsx
@@ -1,4 +1,5 @@
 import { PageLayout } from '@filecoin-foundation/ui/PageLayout'
+import { StructuredDataScript } from '@filecoin-foundation/ui/StructuredDataScript'
 import type { AsyncQueryParams } from '@filecoin-foundation/utils/types/urlTypes'
 
 import { PATHS } from '@/constants/paths'
@@ -14,7 +15,6 @@ import { PageFrontmatterSchema } from '@/schemas/PageFrontmatterSchema'
 import { CTASection } from '@/components/CTASection'
 import { PageHeader } from '@/components/PageHeader'
 import { PageSection } from '@/components/PageSection'
-import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import { EcosystemExplorerContent } from './components/EcosystemExplorerContent'
 import { generateStructuredData } from './utils/generateStructuredData'
