+++ apps/ff-site/src/app/ecosystem-explorer/project-form/page.tsx
@@ -1,5 +1,6 @@
 import { DescriptionText } from '@filecoin-foundation/ui/DescriptionText'
 import { PageLayout } from '@filecoin-foundation/ui/PageLayout'
+import { StructuredDataScript } from '@filecoin-foundation/ui/StructuredDataScript'
 import type { AsyncQueryParams } from '@filecoin-foundation/utils/types/urlTypes'
 
 import { PATHS } from '@/constants/paths'
@@ -11,7 +12,6 @@ import { createMetadata } from '@/utils/createMetadata'
 import { PageFrontmatterSchema } from '@/schemas/PageFrontmatterSchema'
 
 import { PageHeader } from '@/components/PageHeader'
-import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import { getGroupedCategoryOptions } from '../utils/getGroupedCategoryOptions'
