+++ apps/ff-site/src/app/filecoin-plus/page.tsx
@@ -1,5 +1,6 @@
 import { Button } from '@filecoin-foundation/ui/Button'
 import { PageLayout } from '@filecoin-foundation/ui/PageLayout'
+import { StructuredDataScript } from '@filecoin-foundation/ui/StructuredDataScript'
 import { ExternalTextLink } from '@filecoin-foundation/ui/TextLink/ExternalTextLink'
 
 import { PATHS } from '@/constants/paths'
@@ -24,7 +25,6 @@ import { CTASection } from '@/components/CTASection'
 import { FocusAreaCard } from '@/components/FocusAreaCard'
 import { PageHeader } from '@/components/PageHeader'
 import { PageSection } from '@/components/PageSection'
-import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import { aboutData } from './data/aboutData'
 import { applicationData } from './data/applicationData'
