+++ packages/ui/package.json
@@ -33,7 +33,8 @@
     "./TagComponents/*": "./src/TagComponents/*.tsx",
     "./TextLink": "./src/TextLink/index.ts",
     "./TextLink/*": "./src/TextLink/*.tsx",
-    "./Pagination": "./src/Pagination/index.ts"
+    "./Pagination": "./src/Pagination/index.ts",
+    "./StructuredDataScript": "./src/StructuredDataScript.tsx"
   },
   "scripts": {
     "lint": "eslint . --fix"
